feat(playground): make the ComponentDefinedInSetup template reactive
